Drivers: input: misc: Add driver touchscreen-buttons to support physically labeled buttons on touch screen surfaces
Adds a driver for supporting permanet, physically labeled, buttons on
touchscreen surfaces. As are common on android phones designed with
android 1.0 to 2.3 in mind. The driver works by attaching to another
input device and mirroring its events, while inserting key events and
filtering for touches that land on the buttons. Buttons are arbitrary
rectangles configurable via dts.

Signed-off-by: Carl Philipp Klemm <carl@uvos.xyz>