ARM: dts: Add basic support for Motorola mapphone tablets
The Motorola mapphone tablets are quite similar to droid4 but with
a different display and peripheral configuration. Let's add a new
motorola-mapphone-tablet-common.dtsi file for the tablets and basic
support for xyboard mz617. Support for similar tablets like xoom2
mz615 and xyboard mz609 can be added easily.

On the mapphone tablets, there's a DSI to LVDS bridge, and the LCD
backlight controller is lp8550 and not lm3532.

Signed-off-by: Tony Lindgren <tony@atomide.com>